Test with ctapipe 0.23

+++ src/ctapipe_io_lst/calibration.py
@@ -262,7 +262,7 @@ class LSTR0Corrections(TelescopeComponent):
                 r1.waveform -= self.offset.tel[tel_id]
 
             broken_pixels = event.mon.tel[tel_id].pixel_status.hardware_failing_pixels
-            dvred_pixels = (event.lst.tel[tel_id].evt.pixel_status & PixelStatus.DVR_STATUS) == 0
+            dvred_pixels = (event.lst.tel[tel_id].evt.pixel_status & np.uint8(PixelStatus.DVR_STATUS)) == 0
             invalid_pixels = broken_pixels | dvred_pixels
 
             if r1.selected_gain_channel is None:
@@ -306,7 +306,7 @@ class LSTR0Corrections(TelescopeComponent):
                 )
 
             broken_pixels = event.mon.tel[tel_id].pixel_status.hardware_failing_pixels
-            dvred_pixels = (event.lst.tel[tel_id].evt.pixel_status & PixelStatus.DVR_STATUS ) == 0
+            dvred_pixels = (event.lst.tel[tel_id].evt.pixel_status & np.uint8(PixelStatus.DVR_STATUS )) == 0
             invalid_pixels = broken_pixels | dvred_pixels
 
             if r1.selected_gain_channel is None:
